Add a static page for pip release links
`pip --find-links` does not follow client-side redirects. To ease
repository migration, this provides a historical snapshot of the release
links as they were at the time of the migration. To get newer releases,
users will have to update their URLs.

This preserves the client-side redirect for those using a browser.