Add a static page for pip release links (#1)
`pip --find-links` does not follow client-side redirects. To ease
repository migration, this provides a historical snapshot of the release
links as they were at the time of the migration. To get newer releases,
users will have to update their URLs.

This preserves the client-side redirect for those using a browser.

Tested:

Demo at

https://gmngeoffrey.github.io/iree-org.github.io/iree/pip-release-links.html
Note that https://gmngeoffrey.github.io/iree-org.github.io/iree/ 404s,
which is the desired behavior, but the js redirect doesn't work in the
fork because the paths are different. Also, this works:

```shell
pip install \
  -f https://gmngeoffrey.github.io/iree-org.github.io/iree/pip-release-links.html \
  --force-reinstall \
  'iree-compiler>=20221121.334'
```